combine import_config
